Better precision in math

+++ Day1/bmi.c
@@ -2,12 +2,12 @@
 #include <math.h>
 int main()
 {
-	float height,weight;
+	double height,weight;
 	fprintf(stdout,"Enter height and weight: ");
-	scanf("%f %f",&height,&weight);
-	float result=weight/(height*height);
-	fprintf(stdout,"BMI old is %f\n",result);
+	scanf("%lf %lf",&height,&weight);
+	double result=weight/(height*height);
+	fprintf(stdout,"BMI old is %lf\n",result);
 	result=1.3*(weight/pow(height,2.5));
-	fprintf(stdout,"BMI new is %f\n",result);
+	fprintf(stdout,"BMI new is %lf\n",result);
 	return 0;
 }
