+++ Day1/task2a.c
@@ -3,10 +3,10 @@
 
 int main()
 {
-   float r,h;
+   double r,h;
    fprintf(stdout,"Enter R and H: ");
-   scanf("%f %f",&r,&h);
-   float result=3.14*r*r*h;
-   fprintf(stdout,"Water in tank is %f liters\n",result);
+   scanf("%lf %lf",&r,&h);
+   double result=3.14*r*r*h;
+   fprintf(stdout,"Water in tank is %lf liters\n",result);
    return 0;
 }
